lib-sters: move all model structs to a top-level module
