--- lib-sters/src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod kbsig;
-pub mod post;
-pub mod user;
